convet db connect snippet to function
updated the code so now on boot the server will try to establish a connection to the db a number of times before failing. this should make it easier to containerize this application.

addition change; authentication to db is set to basic, this fixes the issue where if the database crashes then both the db and server must be restarted in a sequence.